pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "desafio-nlconsulting"
authors = [
    {name = "madruga665",email = "madruga665@hotmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "openai (>=2.31.0,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "python-multipart (>=0.0.25,<0.0.26)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "chardet (>=7.4.1,<8.0.0)",
    "google-genai (>=1.72.0,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
